requirements.txt
yt-dlp>=2024.1.1
Pillow>=10.0.0
pyinstaller>=6.0.0
